+++ tests/test_multiclass_validator.py
@@ -1,5 +1,5 @@
-from sklearn.datasets import make_classification
 from imblearn.over_sampling import SMOTE
+from sklearn.datasets import make_classification
 
 from oversampleqa.validator import validate_multiclass_oversampling